# backend/internal/shared/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Ressist API",
    "version": "1.0.0",
    "description": "Auth API for Ressist"
  },
  "servers": [
    {
      "url": "https://ressist-api.nodryx.com"
    },
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health check (alias livez)",
        "responses": {
          "200": {
            "description": "ok"
          }
        }
      }
    },
    "/livez": {
      "get": {
        "summary": "Liveness probe",
        "responses": {
          "200": {
            "description": "Service is alive"
          }
        }
      }
    },
    "/metrics": {
      "get": {
        "summary": "Prometheus metrics",
        "responses": {
          "200": {
            "description": "metrics"
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "summary": "OpenAPI document",
        "responses": {
          "200": {
            "description": "OpenAPI JSON"
          }
        }
      }
    },
    "/readyz": {
      "get": {
        "summary": "Readiness probe",
        "responses": {
          "200": {
            "description": "Service is ready"
          },
          "503": {
            "description": "Service is not ready"
          }
        }
      }
    },
    "/v1/api-keys": {
      "get": {
        "summary": "List API keys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of API keys"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "summary": "Create API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "description": "Key label (1-64 chars)"
                  },
                  "expires_in_days": {
                    "type": "integer",
                    "description": "Null = never, 1-3650"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created - api_key shown once"
          },
          "400": {
            "description": "Invalid input"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/api-keys/{id}": {
      "delete": {
        "summary": "Revoke API key",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Revoked"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/v1/assignments": {
      "get": {
        "summary": "Get assignments (supports API key)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Assignments"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/assignments/complete": {
      "post": {
        "summary": "Mark assignment complete",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "assignment_id"
                ],
                "properties": {
                  "assignment_id": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "marked complete"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "not found"
          }
        }
      }
    },
    "/v1/auth/google/callback": {
      "get": {
        "summary": "Google OAuth callback",
        "responses": {
          "307": {
            "description": "Redirect to frontend after auth"
          }
        }
      }
    },
    "/v1/auth/google/login": {
      "get": {
        "summary": "Start Google OAuth login",
        "responses": {
          "307": {
            "description": "Redirect to Google OAuth"
          }
        }
      }
    },
    "/v1/auth/logout": {
      "post": {
        "summary": "Logout and revoke refresh token",
        "responses": {
          "204": {
            "description": "Logged out"
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "summary": "Current user profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "summary": "Refresh access token",
        "responses": {
          "200": {
            "description": "Refreshed access token"
          },
          "401": {
            "description": "Invalid refresh token"
          }
        }
      }
    },
    "/v1/auth/sync": {
      "post": {
        "summary": "Sync LMS data after login (mandatory sync)",
        "description": "Performs LMS synchronization after user login. Should be called by frontend after successful login.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Sync completed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "synced": {
                      "type": "boolean",
                      "description": "Whether sync was successful"
                    },
                    "providers": {
                      "type": "array",
                      "items": {
                        "type": "string"
                      },
                      "description": "List of synced providers"
                    },
                    "totalEvents": {
                      "type": "integer",
                      "description": "Total assignments synced"
                    },
                    "newAssignmentsCount": {
                      "type": "integer",
                      "description": "Number of new assignments found"
                    },
                    "newAssignments": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          },
                          "course": {
                            "type": "string"
                          },
                          "deadline": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "source": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/calendar/preview": {
      "get": {
        "summary": "Get calendar preview from cached assignments",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "parameters": [
          {
            "name": "force",
            "in": "query",
            "description": "Force refresh from LMS sources",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort tasks by: deadline_asc (default), deadline_desc, newest, oldest",
            "schema": {
              "type": "string",
              "enum": [
                "deadline_asc",
                "deadline_desc",
                "newest",
                "oldest"
              ],
              "default": "deadline_asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Calendar preview",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "events": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          },
                          "course": {
                            "type": "string"
                          },
                          "class_code": {
                            "type": "string",
                            "description": "Class code extracted from title (e.g., RA, RB, RC)"
                          },
                          "deadline": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "timeRemaining": {
                            "type": "string"
                          },
                          "deadlineDate": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "source": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "sources": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "provider": {
                            "type": "string"
                          },
                          "count": {
                            "type": "integer"
                          },
                          "success": {
                            "type": "boolean"
                          }
                        }
                      }
                    },
                    "total": {
                      "type": "integer"
                    },
                    "successfulSources": {
                      "type": "integer"
                    },
                    "failedSources": {
                      "type": "integer"
                    },
                    "fromCache": {
                      "type": "boolean"
                    },
                    "sortBy": {
                      "type": "string"
                    },
                    "newAssignmentsCount": {
                      "type": "integer",
                      "description": "Number of newly added assignments"
                    },
                    "newAssignments": {
                      "type": "array",
                      "description": "List of newly added assignments for toast notifications",
                      "items": {
                        "type": "object",
                        "properties": {
                          "title": {
                            "type": "string"
                          },
                          "course": {
                            "type": "string"
                          },
                          "deadline": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "source": {
                            "type": "string"
                          }
                        }
                      }
                    },
                    "lastSyncedAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "nextRefreshAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/calendar/test": {
      "post": {
        "summary": "Test selected LMS sources (cache-based)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "sort",
            "in": "query",
            "description": "Sort tasks by: deadline_asc (default), deadline_desc, newest, oldest",
            "schema": {
              "type": "string",
              "enum": [
                "deadline_asc",
                "deadline_desc",
                "newest",
                "oldest"
              ],
              "default": "deadline_asc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Test result"
          }
        }
      }
    },
    "/v1/courses": {
      "get": {
        "summary": "Get courses (supports API key)",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Courses"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/openapi.json": {
      "get": {
        "summary": "OpenAPI document (versioned)",
        "responses": {
          "200": {
            "description": "OpenAPI JSON"
          }
        }
      }
    },
    "/v1/telegram/test-briefing": {
      "post": {
        "summary": "Send Telegram morning briefing",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "sent"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/telegram/test-reminder": {
      "post": {
        "summary": "Send Telegram test reminder",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "sent"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/v1/user": {
      "get": {
        "summary": "Get current user settings/profile",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current user settings",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "email": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "avatar_url": {
                      "type": "string"
                    },
                    "whatsapp_number": {
                      "type": "string"
                    },
                    "whatsapp_enabled": {
                      "type": "boolean"
                    },
                    "telegram_chat_id": {
                      "type": "string"
                    },
                    "telegram_enabled": {
                      "type": "boolean"
                    },
                    "moodle_enabled": {
                      "type": "boolean"
                    },
                    "moodle_calendar_url": {
                      "type": "string"
                    },
                    "google_classroom_enabled": {
                      "type": "boolean"
                    },
                    "google_connected": {
                      "type": "boolean"
                    },
                    "telegram_bot_username": {
                      "type": "string"
                    },
                    "reminder_hours": {
                      "type": "string"
                    },
                    "morning_briefing": {
                      "type": "boolean"
                    },
                    "muted_courses": {
                      "type": "string"
                    },
                    "course_aliases": {
                      "type": "object"
                    },
                    "class_code": {
                      "type": "string",
                      "description": "User's class code for filtering assignments (e.g., RA, RB, RC)"
                    },
                    "created_at": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update current user settings/profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "whatsapp_number": {
                    "type": "string"
                  },
                  "whatsapp_enabled": {
                    "type": "boolean"
                  },
                  "telegram_chat_id": {
                    "type": "string"
                  },
                  "telegram_enabled": {
                    "type": "boolean"
                  },
                  "moodle_enabled": {
                    "type": "boolean"
                  },
                  "moodle_calendar_url": {
                    "type": "string"
                  },
                  "google_classroom_enabled": {
                    "type": "boolean"
                  },
                  "reminder_hours": {
                    "type": "string"
                  },
                  "morning_briefing": {
                    "type": "boolean"
                  },
                  "muted_courses": {
                    "type": "string"
                  },
                  "course_aliases": {
                    "type": "string"
                  },
                  "class_code": {
                    "type": "string",
                    "description": "User's class code for filtering assignments (e.g., RA, RB, RC)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated user settings"
          }
        }
      }
    },
    "/v1/user/course-aliases": {
      "get": {
        "summary": "Get all course aliases for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Course aliases",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "aliases": {
                      "type": "object",
                      "description": "Map of original course names to aliases",
                      "example": {
                        "Mathematics 101": "Kalkulus",
                        "Introduction to Programming": "Pemrograman Dasar"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Add or update a course alias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "original_name",
                  "alias"
                ],
                "properties": {
                  "original_name": {
                    "type": "string",
                    "description": "Original course name from LMS"
                  },
                  "alias": {
                    "type": "string",
                    "description": "Custom alias for the course"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alias added/updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "aliases": {
                      "type": "object",
                      "description": "Updated map of course aliases"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete a course alias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "original_name"
                ],
                "properties": {
                  "original_name": {
                    "type": "string",
                    "description": "Original course name to remove alias"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Alias deleted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "aliases": {
                      "type": "object",
                      "description": "Updated map of course aliases"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/user/google/disconnect": {
      "post": {
        "summary": "Disconnect Google Classroom for current user",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Disconnected"
          }
        }
      }
    },
    "/v1/user/telegram/verify-code": {
      "post": {
        "summary": "Generate Telegram verify code",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "code generated"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}